updated CCC format description. moved PatternEncoder to its own library.
