+++ README.md
@@ -204,7 +204,7 @@ The supported pattern characters are:
      * C outputs part of the currency code e.g. USD. You can specify 1,2 or 3 C's. Specifying CCC will output the full code regardless of its length.
          * C - U
          * CC - US
-         * CCC - USD
+         * CCC - USD - outputs the full currency code regardless of length.
      * &#35; denotes a digit.
      * 0 denotes a digit and with the addition of defining leading and trailing zeros.
      * , (comma) a placeholder for the grouping separtor
